LayoutTests/platform/gtk/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2.tentative-expected.txt

Button1  Inside1  Nested Invoker 2  Inside2 Button2  Invoker1  Button3  Inside3  Nested Invoker 3 Non-focusable popover Button4
Invoker      after
Show popover Hide popover

FAIL Popover focus navigation assert_equals: Hidden popover should be skipped expected Element node <button id="button2">Button2</button> but got Element node <button id="circular2" popovershowtarget="popover4"></but...
FAIL Circular reference tab navigation promise_test: Unhandled rejection with value: object "InvalidStateError: Element has unexpected visibility state"
FAIL Popover focus returns when popover is hidden by invoker promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
